Fix update interview date in interviewing modal and make helper for extracting date and time
