+++ src/components/JobModal/InterviewingView/InterviewingView.tsx
@@ -1,55 +1,25 @@
 import React, { useState, useEffect } from "react";
 
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch } from "react-redux";
 import { useUpdateJobMutation } from "../../../app/services/job-api";
 import { toggleInterviewedModal } from "../../../features/jobSlice";
-import { RootState } from "../../../app/store";
 
 import styles from "./InterviewingView.module.css";
 import { Box, Typography, TextField, Button } from "@mui/material";
 
-import { useGetAJob } from "../../../hooks/get-a-job";
 import LoadingAnimation from "../../LoadingAnimation";
+
+import { useGetAJob } from "../../../hooks/get-a-job";
+import { extractDate, extractTime } from "../../../helper/date-extractor";
 
 const InterviewingView = () => {
   const { selectedJob, skills } = useGetAJob();
 
   const [updateJob] = useUpdateJobMutation();
   const dispatch = useDispatch();
-  const selectedJobState = useSelector(
-    (state: RootState) => state.job.interviewModal
-  );
 
   const [date, setDate] = useState("");
   const [time, setTime] = useState("");
-
-  const extractDateAndTime = (date: Date) => {
-    const processingDate = date
-      .toLocaleDateString("en-US", {
-        year: "numeric",
-        month: "2-digit",
-        day: "2-digit",
-      })
-      .split("T")[0]
-      .split("/");
-
-    const extractedDate = [
-      processingDate[2],
-      processingDate[0],
-      processingDate[1],
-    ].join("-");
-    setDate(extractedDate);
-    console.log(extractedDate);
-
-    let hours = String(date.getHours());
-    let minutes = String(date.getMinutes());
-    const extractedTime = `${hours.padStart(2, "0")}:${minutes.padStart(
-      2,
-      "0"
-    )}`;
-
-    setTime(extractedTime);
-  };
 
   useEffect(() => {
     if (selectedJob) {
@@ -58,15 +28,16 @@ const InterviewingView = () => {
         date = new Date(selectedJob.interviewDate ?? "");
       }
 
-      extractDateAndTime(date);
+      setDate(extractDate(date));
+      setTime(extractTime(date));
     }
   }, [selectedJob]);
 
   const handleSubmit = () => {
     const dateTime = new Date(date + "T" + time + ":00");
     updateJob({
-      jobId: selectedJobState?.jobCategoryId?.jobId,
-      categoryId: selectedJobState?.jobCategoryId?.categoryId,
+      jobId: selectedJob.job.id,
+      categoryId: selectedJob.category.id,
       interviewDate: dateTime.toISOString(),
       type: "update",
     });
